split into an inventory repository and add quantity to products
